Separate type-guard module
