+++ src/utils/type-guard.ts
@@ -1,0 +1,68 @@
+import {
+  VisualizerElement,
+  VisualizerFreeDrawElement,
+  VisualizerGenericElement,
+  VisualizerLinearElement,
+  VisualizerTextElement,
+} from "../machines/visualizerMachine";
+
+const SHAPE_TYPES: Record<
+  VisualizerElement["shape"],
+  "generic" | "linear" | "freedraw" | "text"
+> = {
+  selection: "generic",
+  rectangle: "generic",
+  ellipse: "generic",
+  line: "linear",
+  arrow: "linear",
+  freedraw: "freedraw",
+  text: "text",
+};
+
+export const isGenericElementShape = (
+  shape: VisualizerElement["shape"]
+): shape is VisualizerGenericElement["shape"] => {
+  return SHAPE_TYPES[shape] === "generic";
+};
+
+export const isGenericElement = (
+  element: VisualizerElement
+): element is VisualizerGenericElement => {
+  return isGenericElementShape(element.shape);
+};
+
+export const isLinearElementShape = (
+  shape: VisualizerElement["shape"]
+): shape is VisualizerLinearElement["shape"] => {
+  return SHAPE_TYPES[shape] === "linear";
+};
+
+export const isLinearElement = (
+  element: VisualizerElement
+): element is VisualizerLinearElement => {
+  return isLinearElementShape(element.shape);
+};
+
+export const isFreeDrawElementShape = (
+  shape: VisualizerElement["shape"]
+): shape is VisualizerFreeDrawElement["shape"] => {
+  return SHAPE_TYPES[shape] === "freedraw";
+};
+
+export const isFreeDrawElement = (
+  element: VisualizerElement
+): element is VisualizerFreeDrawElement => {
+  return isFreeDrawElementShape(element.shape);
+};
+
+export const isTextElementShape = (
+  shape: VisualizerElement["shape"]
+): shape is VisualizerTextElement["shape"] => {
+  return SHAPE_TYPES[shape] === "text";
+};
+
+export const isTextElement = (
+  element: VisualizerElement
+): element is VisualizerTextElement => {
+  return isTextElementShape(element.shape);
+};
